Updated .gitignore to reflect ~/ contents.
Allows me to check out the dotfiles repo directly into my home dir, so I can
more easily and consistently track changes to config files w/out having to
compare them to a separate 'dotfiles' dir. Subdirs to be ignored are called out
explicitly, not wildcarded, so if I add a new subdir, it shows up with git status, so
I can decide how to deal with it.